+++ CHANGES.md
@@ -3,10 +3,10 @@
 ## Unreleased - 2021-xx-xx
 ### Changed
 * Change compression algorithm features flags. [#2250]
-* Deprecate `App::data` and `App::data_factory`. [#????]
+* Deprecate `App::data` and `App::data_factory`. [#2271]
 
 [#2250]: https://github.com/actix/actix-web/pull/2250
-[#????]: https://github.com/actix/actix-web/pull/????
+[#2271]: https://github.com/actix/actix-web/pull/2271
 
 
 ## 4.0.0-beta.7 - 2021-06-17